services/notification-ops-worker/requirements.txt
# VenueIQ — Pinned Production Requirements
# Generated with: pip-compile --generate-hashes requirements.in
# Regenerate:     pip-compile --upgrade --generate-hashes requirements.in
#
# Core services
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-multipart==0.0.12

# GCP SDKs
google-cloud-pubsub==2.23.0
google-cloud-firestore==2.19.0
google-cloud-secret-manager==2.20.0
google-auth==2.34.0
google-auth-httplib2==0.2.0

# HTTP client (notification worker)
httpx==0.27.2

# Security
cryptography==43.0.1

# ML service only (install separately for ml-service image)
# google-cloud-aiplatform==1.66.0
# scikit-learn==1.5.2
# numpy==2.1.1
# joblib==1.4.2
